Rename database package to repository.
